+++ RemGame/Content/Content.mgcb
@@ -13,42 +13,6 @@
 
 #---------------------------------- Content ---------------------------------#
 
-#begin Figures/Level1/Principal/Anim/Principal_Ranged_Chalk.png
-/importer:TextureImporter
-/processor:TextureProcessor
-/processorParam:ColorKeyColor=255,0,255,255
-/processorParam:ColorKeyEnabled=True
-/processorParam:GenerateMipmaps=False
-/processorParam:PremultiplyAlpha=True
-/processorParam:ResizeToPowerOfTwo=False
-/processorParam:MakeSquare=False
-/processorParam:TextureFormat=Color
-/build:Figures/Level1/Principal/Anim/Principal_Ranged_Chalk.png
-
-#begin Figures/Level1/Principal/Anim/Principal_Stand.png
-/importer:TextureImporter
-/processor:TextureProcessor
-/processorParam:ColorKeyColor=255,0,255,255
-/processorParam:ColorKeyEnabled=True
-/processorParam:GenerateMipmaps=False
-/processorParam:PremultiplyAlpha=True
-/processorParam:ResizeToPowerOfTwo=False
-/processorParam:MakeSquare=False
-/processorParam:TextureFormat=Color
-/build:Figures/Level1/Principal/Anim/Principal_Stand.png
-
-#begin Figures/Level1/Principal/Anim/Principal_Walk.png
-/importer:TextureImporter
-/processor:TextureProcessor
-/processorParam:ColorKeyColor=255,0,255,255
-/processorParam:ColorKeyEnabled=True
-/processorParam:GenerateMipmaps=False
-/processorParam:PremultiplyAlpha=True
-/processorParam:ResizeToPowerOfTwo=False
-/processorParam:MakeSquare=False
-/processorParam:TextureFormat=Color
-/build:Figures/Level1/Principal/Anim/Principal_Walk.png
-
 #begin Fonts/Arial.spritefont
 /importer:FontDescriptionImporter
 /processor:FontDescriptionProcessor
@@ -147,7 +111,7 @@
 /processorParam:TextureFormat=Color
 /build:Layers/f-02_front.png
 
-#begin Layers/level/Hands_temp.png
+#begin Layers/level/closingWall.jpg
 /importer:TextureImporter
 /processor:TextureProcessor
 /processorParam:ColorKeyColor=255,0,255,255
@@ -157,7 +121,7 @@
 /processorParam:ResizeToPowerOfTwo=False
 /processorParam:MakeSquare=False
 /processorParam:TextureFormat=Color
-/build:Layers/level/Hands_temp.png;Layers/level/closingWall.png
+/build:Layers/level/closingWall.jpg
 
 #begin Layers/level/level_01_A.jpg
 /importer:TextureImporter
