+++ RemGame/LevelDesgin/Obstacle.cs
@@ -1,5 +1,13 @@
-﻿using Microsoft.Xna.Framework;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
+using Microsoft.Xna.Framework.Input;
+
 using FarseerPhysics.Dynamics;
 using FarseerPhysics.Factories;
