+++ RemGame/LevelDesgin/Table.cs
@@ -1,7 +1,16 @@
-﻿using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
-using FarseerPhysics.Dynamics;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
+
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
+using Microsoft.Xna.Framework.Input;
+
+using FarseerPhysics.Dynamics;
+using FarseerPhysics.Factories;
 namespace RemGame
 {
     class Table : Obstacle
