+++ RemGame/LevelDesgin/Tile.cs
@@ -1,7 +1,12 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
+using Microsoft.Xna.Framework.Input;
+
+using MonoGame.Extended;
 using FarseerPhysics.Dynamics;
 using FarseerPhysics.Factories;
+using System.Collections.Generic;
+using System;
 using Microsoft.Xna.Framework.Content;
 
 namespace RemGame
