+++ RemGame/RemGame.csproj
@@ -41,8 +41,6 @@
     <ApplicationManifest>app.manifest</ApplicationManifest>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Figures\groundedEnemy.cs" />
-    <Compile Include="Figures\Principal.cs" />
     <Compile Include="Figures\Shot.cs" />
     <Compile Include="LevelDesgin\Bag.cs" />
     <Compile Include="LevelDesgin\Chair.cs" />
