+++ RemGame/States/MissionCompleteState.cs
@@ -3,6 +3,7 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
 
+using XELibrary;
 
 namespace RemGame
 {
