+++ RemGame/Utils/PathFinder.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text;
 using Microsoft.Xna.Framework;
 
 namespace RemGame
@@ -18,7 +19,8 @@ namespace RemGame
         private const int CostGoDownPlatform = 5;
         private const int CostGoUpPlatform = 5;
 
-        
+
+
         private static List<PathNode> openList = new List<PathNode>();
 
         private static Dictionary<Vector2, float> nodeCosts =
@@ -50,131 +52,51 @@ namespace RemGame
             List<PathNode> adjacentNodes = new List<PathNode>();
 
             int X = currentNode.GridX;
-            int Y = currentNode.GridY;
-
-
-            if ((Y > 0) && (!pathMap.isPassable(X + 1, Y)))
-            {
-                int d = 1;
-                for (int i = 2; i < 5; i++)//adjust to jump force for each enemey
-                {
-                    if (!pathMap.isPassable(X + 1, Y - i))
-                    {
-                        d++;
-                    }
-                    else
-                        break;
-
-                }
-
-                adjacentNodes.Add(new PathNode(
-                    currentNode,
-                    endNode,
-                    new Vector2(X, Y - d),
-                    CostGoUpPlatform + currentNode.DirectCost, heuristicCalc));
-
-            }
-
-            //progress
-            if ((Y > 0) && (!pathMap.isPassable(X - 1, Y)))
-            {
-                int d = 1;
-                for (int i = 2; i < 5; i++)//adjust to jump force for each enemey
-                {
-                    if (!pathMap.isPassable(X - 1, Y - i))
-                    {
-                        d++;
-                    }
-                    else
-                        break;
-
-                }
-
-
-                adjacentNodes.Add(new PathNode(
-                    currentNode,
-                    endNode,
-                    new Vector2(X, Y - d),
-                    CostGoUpPlatform + currentNode.DirectCost, heuristicCalc));
-
-            }
-
-
-
-            if ((Y > 0) && (pathMap.isPassable(X + 1, Y + 1)) && (!pathMap.isPassable(X, Y + 1)))
-            {
-                int d = 1;
-                for (int i = 2; i < 5; i++)
-                {
-                    if (pathMap.isPassable(X + 1, Y + i))
-                    {
-                        d++;
-                    }
-                    else
-                        break;
-
-                }
-                adjacentNodes.Add(new PathNode(
-                    currentNode,
-                    endNode,
-                    new Vector2(X + 1, Y + d),
-                    CostGoDownPlatform + currentNode.DirectCost, heuristicCalc));
-                
-            }
-
-            //progress
-            if ((Y > 0) && (pathMap.isPassable(X - 1, Y + 1)) && (!pathMap.isPassable(X, Y + 1)))
-            {
-                int d = 1;
-                for (int i = 2; i < 5; i++)
-                {
-                    if (pathMap.isPassable(X - 1, Y + i))
-                    {
-                        d++;
-                    }
-                    else
-                        break;
-
-                }
-                adjacentNodes.Add(new PathNode(
-                    currentNode,
-                    endNode,
-                    new Vector2(X - 1, Y + d),
-                    CostGoDownPlatform + currentNode.DirectCost, heuristicCalc));
-
-            }
-
+            int Y = currentNode.GridY;          
 
             if ((X > 0) && (pathMap.isPassable(X - 1, Y)))
             {
-                if (!pathMap.isPassable(X - 1, Y + 1))
-                {
-                    adjacentNodes.Add(new PathNode(
+                adjacentNodes.Add(new PathNode(
                         currentNode,
                         endNode,
                         new Vector2(X - 1, Y),
                         CostStraight + currentNode.DirectCost, heuristicCalc));
-                }
             }
         
 
             if ((X > 0) && (pathMap.isPassable(X + 1, Y)))
             {
-                if (!pathMap.isPassable(X + 1, Y + 1))
-                {
-
-                    // if(pathMap.isPassable(X + 1, Y))
-                    adjacentNodes.Add(new PathNode(
-                            currentNode,
-                            endNode,
-                            new Vector2(X + 1, Y),
-                            CostStraight + currentNode.DirectCost, heuristicCalc));
-                }
+               // if(pathMap.isPassable(X + 1, Y))
+                adjacentNodes.Add(new PathNode(
+                        currentNode,
+                        endNode,
+                        new Vector2(X + 1, Y),
+                        CostStraight + currentNode.DirectCost, heuristicCalc));
                 
+
             }
             
 
-            
+            if ((Y > 0) && (!pathMap.isPassable(X+1, Y)))
+            {
+                adjacentNodes.Add(new PathNode(
+                    currentNode,
+                    endNode,
+                    new Vector2(X, Y - 1),
+                    CostGoUpPlatform + currentNode.DirectCost, heuristicCalc));
+                
+            }
+
+
+            if ((Y > 0 ) && (pathMap.isPassable(X + 1, Y + 1)) && (!pathMap.isPassable(X, Y + 1)))
+            {
+                adjacentNodes.Add(new PathNode(
+                    currentNode,
+                    endNode,
+                    new Vector2(X+1 , Y + 1),
+                    CostGoDownPlatform + currentNode.DirectCost, heuristicCalc));
+
+            }
 
             return adjacentNodes;
         }
