+++ RemGame/Utils/PhysicsView.cs
@@ -1,7 +1,11 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using FarseerPhysics.Dynamics;
+using Microsoft.Xna.Framework.Input;
 
+using FarseerPhysics.Dynamics;
+using FarseerPhysics.Factories;
+using System.Collections.Generic;
+using System;
 
 namespace RemGame
 {
